backend/pyproject.toml
[project]
name = "portfolio-manager-backend"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.12"
dependencies = [
    "alembic>=1.14.0",
    "fastapi>=0.115.6",
    "jose>=1.0.0",
    "matplotlib>=3.10.6",
    "passlib[bcrypt]>=1.7.4",
    "polars>=1.20.0",
    "prefect>=3.4.1",
    "psycopg2-binary==2.9.10",
    "pydantic-settings>=2.7.1",
    "pyjwt>=2.10.1",
    "python-dotenv>=1.0.1",
    "python-jose>=3.3.0",
    "requests>=2.32.3",
    "scipy>=1.16.2",
    "sqlalchemy>=2.0.37",
    "uvicorn>=0.34.0",
    "yfinance==0.2.58",
]

[tool.ruff.format]
quote-style = "double"
indent-style = "space"
line-ending = "lf"

[tool.uv]
dev-dependencies = [
    "ruff>=0.9.6",
]
